requirements.txt
# 🏗️ Phase 1 : Initialisation & Ingestion de Données
# 1.1 Fondations & Core
numpy>=1.26.0
python-dotenv>=1.0.0
requests>=2.31.0
huggingface-hub>=0.20.0
datasets>=2.17.0

# 1.2 Structure du Code (Squelette)
pydantic>=2.8.0

# 1.3 Ingestion de Documents (Universal Loader)
docling>=1.0.0

# 1.4 Extraction Web (Connecteur Internet)
trafilatura>=1.12.0
beautifulsoup4>=4.12.0

# 1.5 Normalisation (Hygiène des Données)
ftfy>=6.2.0
clean-text>=0.6.0
unidecode>=1.3.0

# 1.6 Nettoyage Linguistique
spacy>=3.7.0


# ⚙️ Phase 2 : Traitement, Indexation & Stockage
# 2.2 Vectorisation (Embeddings)
fastembed>=0.3.0

# 2.3 Stockage Vectoriel & Hybride
lancedb>=0.11.0


# 🔍 Phase 3 : Pipeline de Recherche
# 3.1 Cache Sémantique
gptcache>=0.1.43

# 3.3 Reranking (Sélection Finale)
sentence-transformers>=3.0.0


# 🧠 Phase 4 : Intelligence & Génération
# 4.1 Optimisation Matérielle + LLM CPU
llama-cpp-python>=0.2.90


# 🛡️ Phase 5 : Contrôle Qualité & Maintenance
# 5.1 Évaluation Continue
ragas>=0.1.10


# 🖥️ Phase 6 : Interface Utilisateur
streamlit>=1.31.0
